Move table data, implement pagination and selector country number
Refactor the project by moving the table data to a separate component folder to simplify the codebase. Remove any unnecessary files. Implement pagination for the table data and add a selector to allow users to choose the number of countries displayed per page with also adjust style to make it more easy to read.